add SingUp component
