+++ src/сomponents/UserLogin/Login.module.css
@@ -13,15 +13,7 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  padding: 94px 0 152px;
-}
-
-.login_title {
-  text-transform: uppercase;
-  color: var(--bold-text);
-  font-size: 40px;
-  line-height: 18px;
-  margin: 25px 0;
+  padding: 44px 0 112px;
 }
 
 .fields_container {
