+++ src/сomponents/UserLogin/singInUser/SingInUser.module.css
@@ -1,0 +1,10 @@
+.login_title {
+  text-transform: uppercase;
+  color: var(--bold-text);
+  font-size: 40px;
+  line-height: 18px;
+  margin: 25px 0;
+}
+.btn_sing_in {
+  display: flex;
+}
