+++ src/сomponents/UserLogin/singInUser/interface/ISingInUser.ts
@@ -1,0 +1,12 @@
+export interface ISignInUser {
+    email: string;
+    password: string;
+    isLogin: boolean;
+    secretQuestion?: string;
+}
+export const initialSingInUser = {
+    email: "",
+    password: "",
+    isLogin: false,
+    secretQuestion: "",
+}
